Added refreshOptions method
